Added info about Hacksoc's Google Calendar
Created both a page, which can be a static reference that people will
be able to easily remember and link. It can also be used in future
years, just have the content changes.

We provide an embedded copy of the calendar for the at-a-glance
reference, and for those who haven't yet integrated it into their own
calendar. There are also links to the setup instructions for Google and
Apple calendars, and a note to contact us if they're not using either
clients.

As we've added a page, there is now a direct link from the header menu
to the calendar, so there's a very quick and easy way to access it from
whichever page we're on.