Made CSS for `code` more obvious to users
Code is now put on a grey background to make it more obvious that we're
looking at a code segment - partly also because the monospace font we're
using is very similar to our normal one.

We also make the `line-height` a bit bigger so that we don't get
under/overlap between lines of code.